protecting routes backend and frontend side

+++ frontend/src/components/WorkoutDetails.js
@@ -1,16 +1,21 @@
 import { useWorkoutsContext } from '../hooks/useWorkoutsContext';
-import React from 'react';
+import { useAuthContext } from '../hooks/useAuthContext';
 
-// import date-fns 
+// import date-fns
 
 import formatDistanceToNow from 'date-fns/formatDistanceToNow';
 
 export const WorkoutDetails = ({ workout }) => {
   const { dispatch } = useWorkoutsContext();
-
+  const { user } = useAuthContext();
   const handleClick = async () => {
+    if (!user) {
+      return;
+    }
+
     const response = await fetch('api/workouts/' + workout._id, {
       method: 'DELETE',
+      headers: { Authorization: `Bearer ${user.token}` },
     });
 
     const json = await response.json();
@@ -28,8 +33,12 @@ export const WorkoutDetails = ({ workout }) => {
       <p>
         <strong>Reps:</strong> {workout.reps}
       </p>
-      <p>{formatDistanceToNow(new Date(workout.createdAt), {addSuffix:true})}</p>
-      <span className='material-symbols-outlined' onClick={handleClick}>delete</span>
+      <p>
+        {formatDistanceToNow(new Date(workout.createdAt), { addSuffix: true })}
+      </p>
+      <span className="material-symbols-outlined" onClick={handleClick}>
+        delete
+      </span>
     </div>
   );
 };
